Add block examples to Cargo.toml and update explorer url
