+++ documentation/docs/libraries/rust/how_to/create_block.mdx
@@ -4,12 +4,15 @@ title: How To Create A Block
 
 import CodeBlock from '@theme/CodeBlock';
 import MoreExamples from "./_more_examples.md";
+import SimpleBlock from '!!raw-loader!./../../../../../examples/block/simple_block.rs';
 
 <CodeBlock className="language-rust">
+    {SimpleBlock}
 </CodeBlock>
 
 Run it with
 <CodeBlock className="language-bash">
+    cargo run --example simple_block --release
 </CodeBlock>
 
 <MoreExamples/>
