+++ examples/03_simple_block.rs
@@ -18,13 +18,14 @@ async fn main() -> Result<()> {
 
     let client = Client::builder()
         .with_node(&node_url)?
+        .with_pow_worker_count(1)
         .with_node_sync_disabled()
         .finish()?;
 
     let block = client.block().finish().await?;
 
     println!(
-        "Empty block sent: https://explorer.iota.org/devnet/block/{}",
+        "Empty block sent: https://explorer.testnet.shimmer.network/testnet/block/{}",
         block.id()
     );
     Ok(())
