+++ examples/block/custom_inputs.rs
@@ -59,7 +59,7 @@ async fn main() -> Result<()> {
         .await?;
 
     println!(
-        "Transaction sent: https://explorer.iota.org/devnet/block/{}",
+        "Transaction sent: https://explorer.testnet.shimmer.network/testnet/block/{}",
         block.id()
     );
     Ok(())
