+++ examples/block/data_block.rs
@@ -1,7 +1,7 @@
 // Copyright 2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-//! cargo run --example 05_data_block --release
+//! cargo run --example data_block --release
 
 use std::env;
 
@@ -18,23 +18,29 @@ async fn main() -> Result<()> {
 
     let client = Client::builder()
         .with_node(&node_url)?
-        // .with_permanode("http://18.196.167.57:8000/api/permanode/", None, None)?
         .with_node_sync_disabled()
         .finish()?;
 
     let block = client
         .block()
-        .with_tag("Hello")
+        .with_tag("Hello".as_bytes().to_vec())
         .with_data("Tangle".as_bytes().to_vec())
         .finish()
         .await?;
 
-    println!("Block sent https://explorer.iota.org/devnet/block/{}\n", block.id());
+    println!(
+        "Block sent https://explorer.testnet.shimmer.network/testnet/block/{}\n",
+        block.id()
+    );
 
     let fetched_block = client.get_block(&block.id()).await?;
     println!("{:#?}\n", fetched_block);
 
-    if let Payload::TaggedData(payload) = fetched_block.payload().as_ref()? {
+    if let Some(Payload::TaggedData(payload)) = fetched_block.payload() {
+        println!(
+            "Tag: {}",
+            String::from_utf8(payload.tag().to_vec()).expect("Found invalid UTF-8")
+        );
         println!(
             "Data: {}",
             String::from_utf8(payload.data().to_vec()).expect("Found invalid UTF-8")
