+++ examples/custom_remainder_address.rs
@@ -68,7 +68,7 @@ async fn main() -> Result<()> {
         .await?;
 
     println!(
-        "Transaction sent: https://explorer.iota.org/devnet/block/{}",
+        "Transaction sent: https://explorer.testnet.shimmer.network/testnet/block/{}",
         block.id()
     );
